move exported functions to a separate file
